# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Open Music Player API
  description: |
    API for the Open Music Player application.
    Provides endpoints for authentication, music search (via MusicBrainz),
    library management, playlists, downloads, signed playback URLs, queue management,
    and explicit YouTube playlist imports.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.openmusicplayer.app/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login with credentials
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Logout and invalidate tokens
      operationId: logoutUser
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # Search Endpoints
  # ============================================================================
  /search/tracks:
    get:
      tags:
        - Search
      summary: Search for tracks on MusicBrainz
      operationId: searchTracks
      parameters:
        - $ref: '#/components/parameters/QueryParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /search/artists:
    get:
      tags:
        - Search
      summary: Search for artists on MusicBrainz
      operationId: searchArtists
      parameters:
        - $ref: '#/components/parameters/QueryParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /search/albums:
    get:
      tags:
        - Search
      summary: Search for albums on MusicBrainz
      operationId: searchAlbums
      parameters:
        - $ref: '#/components/parameters/QueryParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # Browse/Discovery Endpoints
  # ============================================================================
  /artists/{mb_id}:
    get:
      tags:
        - Browse
      summary: Get artist details with discography
      operationId: getArtist
      parameters:
        - $ref: '#/components/parameters/MusicBrainzIdParam'
      responses:
        '200':
          description: Artist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /albums/{mb_id}:
    get:
      tags:
        - Browse
      summary: Get album details with track listing
      operationId: getAlbum
      parameters:
        - $ref: '#/components/parameters/MusicBrainzIdParam'
      responses:
        '200':
          description: Album details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /tracks/{mb_id}:
    get:
      tags:
        - Browse
      summary: Get track details
      operationId: getTrack
      parameters:
        - $ref: '#/components/parameters/MusicBrainzIdParam'
      responses:
        '200':
          description: Track details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /discovery/search:
    get:
      tags: [Browse]
      summary: Search ranked source candidates and create a short-lived selection session
      operationId: discoverySearch
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: providers
          in: query
          schema: { type: string, description: Comma-separated provider names. }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 25 }
      responses:
        '200':
          description: Ranked candidates; selectionRequired is false only when results is empty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscoverySearchResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /discovery/resolve-url:
    post:
      tags: [Browse]
      summary: Resolve a pasted supported URL and create a short-lived selection session
      operationId: resolveDiscoveryURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [url]
              properties: { url: { type: string, format: uri } }
      responses:
        '200':
          description: Resolved candidate and selection session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResolveDiscoveryURLResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /discovery/assist:
    post:
      tags: [Browse]
      summary: Ground an assistant request in discovered candidates and create a selection session when candidates are returned
      operationId: assistDiscovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [prompt]
              properties:
                prompt: { type: string, minLength: 1 }
                limit: { type: integer, minimum: 1, maximum: 25 }
      responses:
        '200':
          description: Grounded assistant response. Selection metadata is top-level for direct and nested search candidates.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscoveryAssistResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ============================================================================
  # Library Endpoints
  # ============================================================================
  /library/tracks:
    get:
      tags:
        - Library
      summary: Get user's track library
      operationId: getLibraryTracks
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, artist, album, added_at, duration]
            default: added_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of tracks in library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Library
      summary: Add a track to library
      operationId: addTrackToLibrary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTrackRequest'
      responses:
        '201':
          description: Track added to library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Track already in library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /library/tracks/{trackId}:
    get:
      tags:
        - Library
      summary: Get a specific track from library
      operationId: getLibraryTrack
      parameters:
        - $ref: '#/components/parameters/TrackIdParam'
      responses:
        '200':
          description: Track details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Library
      summary: Remove a track from library
      operationId: removeTrackFromLibrary
      parameters:
        - $ref: '#/components/parameters/TrackIdParam'
      responses:
        '204':
          description: Track removed from library
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # Playlist Endpoints
  # ============================================================================
  /playlists:
    get:
      tags:
        - Playlists
      summary: Get user's playlists
      operationId: getPlaylists
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: List of playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Playlists
      summary: Create a new playlist
      operationId: createPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /playlists/{playlistId}:
    get:
      tags:
        - Playlists
      summary: Get playlist details
      operationId: getPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistIdParam'
      responses:
        '200':
          description: Playlist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistWithTracks'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Playlists
      summary: Update playlist details
      operationId: updatePlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistRequest'
      responses:
        '200':
          description: Playlist updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Playlists
      summary: Delete a playlist
      operationId: deletePlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistIdParam'
      responses:
        '204':
          description: Playlist deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /playlists/{playlistId}/tracks/{trackId}:
    delete:
      tags:
        - Playlists
      summary: Remove a track from playlist
      operationId: removeTrackFromPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistIdParam'
        - $ref: '#/components/parameters/TrackIdParam'
      responses:
        '204':
          description: Track removed from playlist
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /downloads/{job_id}:
    get:
      tags:
        - Downloads
      summary: Get download job status
      operationId: getDownloadStatus
      parameters:
        - $ref: '#/components/parameters/DownloadJobIdParam'
      responses:
        '200':
          description: Download job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        username:
          type: string
          minLength: 3
          maxLength: 50
...
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'

    # ========================================================================
    # User Schema
    # ========================================================================
...