# backend/internal/api/router.go
package api

import (
	"encoding/json"
	"net/http"

	"github.com/openmusicplayer/backend/internal/auth"
	"github.com/openmusicplayer/backend/internal/discovery"
	apperrors "github.com/openmusicplayer/backend/internal/errors"
	"github.com/openmusicplayer/backend/internal/health"
	"github.com/openmusicplayer/backend/internal/logger"
	"github.com/openmusicplayer/backend/internal/matcher"
	"github.com/openmusicplayer/backend/internal/metrics"
	"github.com/openmusicplayer/backend/internal/middleware"
	"github.com/openmusicplayer/backend/internal/musicbrainz"
	"github.com/openmusicplayer/backend/internal/queue"
	"github.com/openmusicplayer/backend/internal/search"
	"github.com/openmusicplayer/backend/internal/validators"
	"github.com/openmusicplayer/backend/internal/websocket"
)

type Router struct {
	mux                     *http.ServeMux
	authHandlers            *auth.Handlers
	authService             *auth.Service
	searchHandlers          *search.Handlers
	browseHandlers          *BrowseHandlers
	musicbrainzHandlers     *musicbrainz.Handlers
	wsHandler               *websocket.Handler
	validatorHandlers       *validators.Handlers
	matcherHandlers         *matcher.Handler
	libraryHandlers         *LibraryHandlers
	analysisHandlers        *AnalysisHandlers
	playbackHandlers        *PlaybackHandlers
	queueHandlers           *queue.Handlers
	discoveryHandlers       *discovery.Handlers
	playlistHandlers        *PlaylistHandlers
	playlistImportHandlers  *PlaylistImportHandlers
	playlistMixHandlers     *PlaylistMixHandlers
	mixPlanHandlers         *MixPlanHandlers
	downloadHandlers        *DownloadHandlers
	sourceSelectionHandlers *SourceSelectionHandlers
	maintenanceHandlers     *MaintenanceHandlers
	playEventHandlers       *PlayEventHandlers
	healthHandler           *health.Handler
	metricsHandler          http.HandlerFunc
	corsAllowedOrigins      []string
}

var defaultCORSAllowedOrigins = []string{
	"http://localhost:18145",
	"http://127.0.0.1:18145",
}

// RouterConfig holds configuration for creating a new router
type RouterConfig struct {
	AuthHandlers            *auth.Handlers
	AuthService             *auth.Service
	SearchHandlers          *search.Handlers
	MBClient                *musicbrainz.Client
	MBHandlers              *musicbrainz.Handlers
	WSHandler               *websocket.Handler
	MatcherHandlers         *matcher.Handler
	LibraryHandlers         *LibraryHandlers
	AnalysisHandlers        *AnalysisHandlers
	PlaybackHandlers        *PlaybackHandlers
	QueueHandlers           *queue.Handlers
	DiscoveryHandlers       *discovery.Handlers
	PlaylistHandlers        *PlaylistHandlers
	PlaylistImportHandlers  *PlaylistImportHandlers
	PlaylistMixHandlers     *PlaylistMixHandlers
	MixPlanHandlers         *MixPlanHandlers
	DownloadHandlers        *DownloadHandlers
	SourceSelectionHandlers *SourceSelectionHandlers
	MaintenanceHandlers     *MaintenanceHandlers
	PlayEventHandlers       *PlayEventHandlers
	HealthHandler           *health.Handler
	Metrics                 *metrics.Metrics
	CORSAllowedOrigins      []string
}

func NewRouter(authHandlers *auth.Handlers, authService *auth.Service, searchHandlers *search.Handlers, mbClient *musicbrainz.Client, mbHandlers *musicbrainz.Handlers, wsHandler *websocket.Handler, matcherHandlers *matcher.Handler, libraryHandlers *LibraryHandlers, queueHandlers *queue.Handlers, playlistHandlers *PlaylistHandlers, downloadHandlers *DownloadHandlers) *Router {
	return NewRouterWithConfig(&RouterConfig{
		AuthHandlers:     authHandlers,
		AuthService:      authService,
		SearchHandlers:   searchHandlers,
		MBClient:         mbClient,
		MBHandlers:       mbHandlers,
		WSHandler:        wsHandler,
		MatcherHandlers:  matcherHandlers,
		LibraryHandlers:  libraryHandlers,
		QueueHandlers:    queueHandlers,
		PlaylistHandlers: playlistHandlers,
		DownloadHandlers: downloadHandlers,
	})
}

func NewRouterWithConfig(cfg *RouterConfig) *Router {
	validatorRegistry := validators.DefaultRegistry()

	var metricsHandler http.HandlerFunc
	if cfg.Metrics != nil {
		metricsHandler = cfg.Metrics.Handler()
	}

	corsAllowedOrigins := cfg.CORSAllowedOrigins
	if corsAllowedOrigins == nil {
		corsAllowedOrigins = defaultCORSAllowedOrigins
	}

	r := &Router{
		mux:                     http.NewServeMux(),
		authHandlers:            cfg.AuthHandlers,
		authService:             cfg.AuthService,
		searchHandlers:          cfg.SearchHandlers,
		browseHandlers:          NewBrowseHandlers(cfg.MBClient),
		musicbrainzHandlers:     cfg.MBHandlers,
		wsHandler:               cfg.WSHandler,
		validatorHandlers:       validators.NewHandlers(validatorRegistry),
		matcherHandlers:         cfg.MatcherHandlers,
		libraryHandlers:         cfg.LibraryHandlers,
		analysisHandlers:        cfg.AnalysisHandlers,
		playbackHandlers:        cfg.PlaybackHandlers,
		queueHandlers:           cfg.QueueHandlers,
		discoveryHandlers:       cfg.DiscoveryHandlers,
		playlistHandlers:        cfg.PlaylistHandlers,
		playlistImportHandlers:  cfg.PlaylistImportHandlers,
		playlistMixHandlers:     cfg.PlaylistMixHandlers,
		mixPlanHandlers:         cfg.MixPlanHandlers,
		downloadHandlers:        cfg.DownloadHandlers,
		sourceSelectionHandlers: cfg.SourceSelectionHandlers,
		maintenanceHandlers:     cfg.MaintenanceHandlers,
		playEventHandlers:       cfg.PlayEventHandlers,
		healthHandler:           cfg.HealthHandler,
		metricsHandler:          metricsHandler,
		corsAllowedOrigins:      corsAllowedOrigins,
	}
	r.setupRoutes()
	return r
}

func (r *Router) ServeHTTP(w http.ResponseWriter, req *http.Request) {
	// Apply middleware chain: CORS -> Recovery -> RequestID -> Logging -> Routes
	handler := middleware.CORS(r.corsAllowedOrigins)(
		logger.RecoveryMiddleware(
			apperrors.RequestIDMiddleware(
				logger.LoggingMiddleware(r.mux),
			),
		),
	)
	handler.ServeHTTP(w, req)
}

func (r *Router) setupRoutes() {
	// Health check endpoints (Kubernetes-compatible)
	if r.healthHandler != nil {
		r.mux.HandleFunc("GET /health", r.healthHandler.HealthHandler)
		r.mux.HandleFunc("GET /health/live", r.healthHandler.LivenessHandler)
		r.mux.HandleFunc("GET /health/ready", r.healthHandler.ReadinessHandler)
	} else {
		// Fallback to simple health check if handler not configured
		r.mux.HandleFunc("GET /health", defaultHealthHandler)
	}

	// Metrics endpoint (Prometheus-compatible)
	if r.metricsHandler != nil {
		r.mux.HandleFunc("GET /metrics", r.metricsHandler)
	}

	// Auth routes (no auth required)
	r.mux.HandleFunc("POST /api/v1/auth/register", r.authHandlers.Register)
	r.mux.HandleFunc("POST /api/v1/auth/login", r.authHandlers.Login)
	r.mux.HandleFunc("POST /api/v1/auth/refresh", r.authHandlers.Refresh)

	// Auth routes (auth required)
	r.mux.HandleFunc("POST /api/v1/auth/logout", r.withAuth(r.authHandlers.Logout))

	// Search routes - local database (auth required)
	r.mux.HandleFunc("GET /api/v1/search", r.withAuth(r.searchHandlers.Search))
	r.mux.HandleFunc("GET /api/v1/search/recordings", r.withAuth(r.searchHandlers.SearchRecordings))
	r.mux.HandleFunc("GET /api/v1/search/artists", r.withAuth(r.searchHandlers.SearchArtists))
	r.mux.HandleFunc("GET /api/v1/search/releases", r.withAuth(r.searchHandlers.SearchReleases))

	// Search routes - MusicBrainz with caching (auth required)
	r.mux.HandleFunc("GET /api/v1/musicbrainz/search/tracks", r.withAuth(r.musicbrainzHandlers.SearchTracks))
	r.mux.HandleFunc("GET /api/v1/musicbrainz/search/artists", r.withAuth(r.musicbrainzHandlers.SearchArtists))
	r.mux.HandleFunc("GET /api/v1/musicbrainz/search/albums", r.withAuth(r.musicbrainzHandlers.SearchAlbums))

	// Browse/discovery routes (auth required)
	if r.discoveryHandlers != nil {
		r.mux.HandleFunc("GET /api/v1/discovery/search", r.withAuth(r.discoveryHandlers.Search))
		r.mux.HandleFunc("POST /api/v1/discovery/resolve-url", r.withAuth(r.discoveryHandlers.ResolveURL))
		r.mux.HandleFunc("POST /api/v1/discovery/assist", r.withAuth(r.discoveryHandlers.Assist))
	} else {
		r.mux.HandleFunc("GET /api/v1/discovery/search", r.withAuth(unavailableHandler("Discovery search is unavailable")))
		r.mux.HandleFunc("POST /api/v1/discovery/resolve-url", r.withAuth(unavailableHandler("Discovery URL resolver is unavailable")))
		r.mux.HandleFunc("POST /api/v1/discovery/assist", r.withAuth(unavailableHandler("Discovery assist is unavailable")))
	}
	if r.sourceSelectionHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/source-selections", r.withAuth(r.sourceSelectionHandlers.Create))
		r.mux.HandleFunc("GET /api/v1/source-selections", r.withAuth(r.sourceSelectionHandlers.List))
		r.mux.HandleFunc("GET /api/v1/source-selections/{id}", r.withAuth(r.sourceSelectionHandlers.Get))
	} else {
		sourceSelectionUnavailable := r.withAuth(unavailableHandler("Source selections are unavailable"))
		r.mux.HandleFunc("POST /api/v1/source-selections", sourceSelectionUnavailable)
		r.mux.HandleFunc("GET /api/v1/source-selections", sourceSelectionUnavailable)
		r.mux.HandleFunc("GET /api/v1/source-selections/{id}", sourceSelectionUnavailable)
	}
	r.mux.HandleFunc("GET /api/v1/artists/{mb_id}", r.withAuth(r.browseHandlers.GetArtist))
	r.mux.HandleFunc("GET /api/v1/albums/{mb_id}", r.withAuth(r.browseHandlers.GetAlbum))
	r.mux.HandleFunc("GET /api/v1/tracks/{mb_id}", r.withAuth(r.browseHandlers.GetTrack))

	// WebSocket route (auth via query param)
	r.mux.HandleFunc("GET /api/v1/ws/progress", r.wsHandler.ServeWS)

	// URL validation routes (auth required)
	r.mux.HandleFunc("POST /api/v1/validate/url", r.withAuth(r.validatorHandlers.ValidateURL))
	r.mux.HandleFunc("GET /api/v1/validate/url", r.withAuth(r.validatorHandlers.ValidateURLQuery))
	r.mux.HandleFunc("GET /api/v1/validate/sources", r.withAuth(r.validatorHandlers.GetSupportedSources))

	// Auto-matching routes (auth required)
	r.mux.HandleFunc("POST /api/v1/match", r.withAuth(r.matcherHandlers.HandleMatch))
	r.mux.HandleFunc("POST /api/v1/tracks/{id}/match", r.withAuth(r.matcherHandlers.HandleMatchTrack))
	r.mux.HandleFunc("POST /api/v1/tracks/{id}/confirm-match", r.withAuth(r.matcherHandlers.HandleConfirmMatch))
	r.mux.HandleFunc("POST /api/v1/tracks/{id}/link-mb", r.withAuth(r.matcherHandlers.HandleLinkMB))

	// Library routes (auth required)
	r.mux.HandleFunc("GET /api/v1/library", r.withAuth(r.libraryHandlers.GetLibrary))
	r.mux.HandleFunc("POST /api/v1/library/tracks/{track_id}", r.withAuth(r.libraryHandlers.AddTrackToLibrary))
	r.mux.HandleFunc("DELETE /api/v1/library/tracks/{track_id}", r.withAuth(r.libraryHandlers.RemoveTrackFromLibrary))
	r.mux.HandleFunc("POST /api/v1/library/tracks/{track_id}/like", r.withAuth(r.libraryHandlers.LikeTrack))
	r.mux.HandleFunc("DELETE /api/v1/library/tracks/{track_id}/like", r.withAuth(r.libraryHandlers.UnlikeTrack))
	if r.analysisHandlers != nil {
		r.mux.HandleFunc("GET /api/v1/tracks/{track_id}/analysis", r.withAuth(r.analysisHandlers.GetTrackAnalysis))
		r.mux.HandleFunc("PATCH /api/v1/tracks/{track_id}/analysis/overrides", r.withAuth(r.analysisHandlers.UpdateTrackAnalysisOverrides))
	} else {
		r.mux.HandleFunc("GET /api/v1/tracks/{track_id}/analysis", r.withAuth(unavailableHandler("Track analysis is unavailable")))
		r.mux.HandleFunc("PATCH /api/v1/tracks/{track_id}/analysis/overrides", r.withAuth(unavailableHandler("Track analysis is unavailable")))
	}

	// Direct playback/download URL issuance (auth required)
	if r.playbackHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/playback/urls", r.withAuth(r.playbackHandlers.CreatePlaybackURLs))
	} else {
		r.mux.HandleFunc("POST /api/v1/playback/urls", r.withAuth(unavailableHandler("Playback URL issuance is unavailable")))
	}

	// Queue routes (auth required, Redis-backed)
	if r.queueHandlers != nil {
		r.mux.HandleFunc("GET /api/v1/queue", r.withAuth(r.queueHandlers.GetQueue))
		r.mux.HandleFunc("POST /api/v1/queue/items", r.withAuth(r.queueHandlers.AddQueueItem))
		r.mux.HandleFunc("POST /api/v1/queue/items/{queueItemId}/retry", r.withAuth(r.queueHandlers.RetryQueueItem))
		r.mux.HandleFunc("DELETE /api/v1/queue/items/{queueItemId}", r.withAuth(r.queueHandlers.RemoveQueueItem))
		r.mux.HandleFunc("PUT /api/v1/queue/reorder", r.withAuth(r.queueHandlers.ReorderQueue))
		r.mux.HandleFunc("DELETE /api/v1/queue", r.withAuth(r.queueHandlers.ClearQueue))
	} else {
		queueUnavailable := r.withAuth(unavailableHandler("Redis queue support is disabled for this local mode"))
		r.mux.HandleFunc("GET /api/v1/queue", queueUnavailable)
		r.mux.HandleFunc("POST /api/v1/queue/items", queueUnavailable)
		r.mux.HandleFunc("POST /api/v1/queue/items/{queueItemId}/retry", queueUnavailable)
		r.mux.HandleFunc("DELETE /api/v1/queue/items/{queueItemId}", queueUnavailable)
		r.mux.HandleFunc("PUT /api/v1/queue/reorder", queueUnavailable)
		r.mux.HandleFunc("DELETE /api/v1/queue", queueUnavailable)
	}

	// Playlist routes (auth required)
	r.mux.HandleFunc("GET /api/v1/playlists", r.withAuth(r.playlistHandlers.ListPlaylists))
	r.mux.HandleFunc("POST /api/v1/playlists", r.withAuth(r.playlistHandlers.CreatePlaylist))
	r.mux.HandleFunc("GET /api/v1/playlists/{id}", r.withAuth(r.playlistHandlers.GetPlaylist))
	r.mux.HandleFunc("PUT /api/v1/playlists/{id}", r.withAuth(r.playlistHandlers.UpdatePlaylist))
	r.mux.HandleFunc("DELETE /api/v1/playlists/{id}", r.withAuth(r.playlistHandlers.DeletePlaylist))
	r.mux.HandleFunc("POST /api/v1/playlists/{id}/tracks", r.withAuth(r.playlistHandlers.AddTracks))
	r.mux.HandleFunc("DELETE /api/v1/playlists/{id}/tracks/{trackId}", r.withAuth(r.playlistHandlers.RemoveTrack))
	r.mux.HandleFunc("POST /api/v1/playlists/{id}/tracks/batch-remove", r.withAuth(r.playlistHandlers.BatchRemoveTracks))
	r.mux.HandleFunc("PUT /api/v1/playlists/{id}/tracks/reorder", r.withAuth(r.playlistHandlers.ReorderTracks))
	// Flag-gated save-playlist-as-mix seam. The handler itself returns 404 when
	// the feature is disabled (ENABLE_PLAYLIST_MIX); when the handler is not wired
	// at all (legacy router construction) the route stays unregistered.
	if r.playlistMixHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/playlists/{id}/mix", r.withAuth(r.playlistMixHandlers.CreateMixFromPlaylist))
	}
	if r.playlistImportHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/playlist-imports", r.withAuth(r.playlistImportHandlers.CreateImport))
		r.mux.HandleFunc("GET /api/v1/playlist-imports/{importJobId}", r.withAuth(r.playlistImportHandlers.GetImport))
	} else {
		playlistImportUnavailable := r.withAuth(unavailableHandler("Playlist import processing is disabled for this local mode"))
		r.mux.HandleFunc("POST /api/v1/playlist-imports", playlistImportUnavailable)
		r.mux.HandleFunc("GET /api/v1/playlist-imports/{importJobId}", playlistImportUnavailable)
	}

	// Saved mix plan routes (auth required). The server stores durable plan state only;
	// playback/rendering state stays client-side.
	r.mux.HandleFunc("GET /api/v1/mix-plans", r.withAuth(r.mixPlanHandlers.ListMixPlans))
	r.mux.HandleFunc("POST /api/v1/mix-plans", r.withAuth(r.mixPlanHandlers.CreateMixPlan))
	r.mux.HandleFunc("GET /api/v1/mix-plans/{mixPlanId}", r.withAuth(r.mixPlanHandlers.GetMixPlan))
	r.mux.HandleFunc("PUT /api/v1/mix-plans/{mixPlanId}", r.withAuth(r.mixPlanHandlers.UpdateMixPlan))

	// Download routes (auth required, Redis/worker-backed)
	if r.downloadHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/downloads", r.withAuth(r.downloadHandlers.CreateDownload))
		r.mux.HandleFunc("GET /api/v1/downloads", r.withAuth(r.downloadHandlers.GetUserJobs))
		r.mux.HandleFunc("GET /api/v1/downloads/{job_id}", r.withAuth(r.downloadHandlers.GetJob))
	} else {
		downloadUnavailable := r.withAuth(unavailableHandler("Download processing is disabled for this local mode"))
		r.mux.HandleFunc("POST /api/v1/downloads", downloadUnavailable)
		r.mux.HandleFunc("GET /api/v1/downloads", downloadUnavailable)
		r.mux.HandleFunc("GET /api/v1/downloads/{job_id}", downloadUnavailable)
	}

	// Play event routes (auth required): record a play and read personal history.
	if r.playEventHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/me/plays", r.withAuth(r.playEventHandlers.RecordPlay))
		r.mux.HandleFunc("GET /api/v1/me/plays/history", r.withAuth(r.playEventHandlers.PlayHistory))
		r.mux.HandleFunc("GET /api/v1/me/plays/recent", r.withAuth(r.playEventHandlers.RecentlyPlayed))
		r.mux.HandleFunc("GET /api/v1/me/plays/top", r.withAuth(r.playEventHandlers.TopTracks))
	} else {
		playEventUnavailable := r.withAuth(unavailableHandler("Play history is unavailable"))
		r.mux.HandleFunc("POST /api/v1/me/plays", playEventUnavailable)
		r.mux.HandleFunc("GET /api/v1/me/plays/history", playEventUnavailable)
		r.mux.HandleFunc("GET /api/v1/me/plays/recent", playEventUnavailable)
		r.mux.HandleFunc("GET /api/v1/me/plays/top", playEventUnavailable)
	}

	// Maintenance repair routes (auth required)
	if r.maintenanceHandlers != nil {
		r.mux.HandleFunc("POST /api/v1/maintenance/repair", r.withAuth(r.maintenanceHandlers.RepairTracks))
	} else {
		r.mux.HandleFunc("POST /api/v1/maintenance/repair", r.withAuth(unavailableHandler("Maintenance repair is unavailable")))
	}
}

func unavailableHandler(message string) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusServiceUnavailable)
		json.NewEncoder(w).Encode(map[string]string{
			"code":    "SERVICE_DISABLED",
			"message": message,
		})
	}
}

func (r *Router) withAuth(next http.HandlerFunc) http.HandlerFunc {
	middleware := auth.Middleware(r.authService)
	return func(w http.ResponseWriter, req *http.Request) {
		middleware(next).ServeHTTP(w, req)
	}
}

func defaultHealthHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]string{
		"status": "ok",
	})
}